requirements.txt
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
asyncpg>=0.29,<1
openai>=1.30,<2
gspread>=6.0,<7
python-telegram-bot>=21.0,<22
